Network - Split rpc services into several files
